Expose a little bit of code in the headers to improve inlining.
